# Reduce cost of no attribute counters