Fixes #
Design discussion issue (if applicable) #

## Changes

This PR utilizes atomics to make counters fast when an empty set of attributes are passed in.  This atomic bypasses the hashmap lookups and mutex locking.

## Benchmarks

### main
```
Counter/AddNoAttrs      time:   [59.516 ns 59.947 ns 60.457 ns]
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe

Counter/AddNoAttrsDelta time:   [62.342 ns 63.361 ns 64.458 ns]
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

Counter/AddOneAttr      time:   [144.01 ns 146.76 ns 150.11 ns]
Found 13 outliers among 100 measurements (13.00%)
  5 (5.00%) high mild
  8 (8.00%) high severe

Counter/AddOneAttrDelta time:   [144.58 ns 145.39 ns 146.28 ns]
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) high mild
  5 (5.00%) high severe
```

### PR
```
Counter/AddNoAttrs      time:   [30.140 ns 30.390 ns 30.652 ns]
                        change: [-49.181% -48.265% -47.331%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe
Counter/AddNoAttrsDelta time:   [29.836 ns 30.040 ns 30.327 ns]
                        change: [-53.397% -52.564% -51.790%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
Counter/AddOneAttr      time:   [145.48 ns 145.84 ns 146.24 ns]
                        change: [-0.3877% +0.8455% +2.1208%] (p = 0.20 > 0.05)
                        No change in performance detected.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) high mild
  7 (7.00%) high severe
Counter/AddOneAttrDelta time:   [147.37 ns 148.07 ns 148.85 ns]
                        change: [+0.5418% +2.2081% +3.9805%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 11 outliers among 100 measurements (11.00%)
  4 (4.00%) high mild
  7 (7.00%) high severe
```

## Merge requirement checklist

* [X] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-rust/blob/main/CONTRIBUTING.md) guidelines followed
* [X] Unit tests added/updated (if applicable)
* [X] Appropriate `CHANGELOG.md` files updated for non-trivial, user-facing changes
* [ ] Changes in public API reviewed (if applicable)