@cijothomas 2024-02-14
Consider a changelog entry, as users upgrading will learn about the perf improvement of zero-attributes, and can leverage it to achieve bound-instrument (not technically, but a reasonable workaround).